Delete the determination pass the sweep replaced, and operators::common
`discover_times` lost its last caller when the fused sweep landed.
`KeyView`, `DiscoverScratch` and `TimeHistory` existed only to serve it.
The containment test that checked the sweep against it goes too: it was scaffolding for building the sweep, and the ten `int_proxy` tests and the corgi gate check the same property end to end against the conventional reduce.

`common.rs` said "Types and methods generally useful for differential computation".
Every live item in it had exactly one caller, and all three callers were in `int_proxy`.
Naming it that way claimed a generality it did not have, hid the tactic's size from anyone reading the tactic, and is how four items died unnoticed: a module named for a role nobody audits accumulates things.

So `Sweep`, its `update_meet` helper and `tile_descriptions` move to the tail of `int_proxy/reduce.rs`, and `bilinear_wave` to `int_proxy/join.rs`, in each case beside their only caller.
None of them is `pub` any more; none was ever part of an interface.
The sweep and the tactic that drives it now sit in one file, which is where they belong: the boundary between them is meant to dissolve as more of the work fuses.

A warning for anyone who benchmarks this commit.
Under the workspace release profile it appears to cost 8% on churn and multimoment, and that is an artifact of `codegen-units`, which the profile leaves at the default 16 despite `lto = true`.
Rustc partitions codegen units by module, so folding two modules into one merges two partitions and the optimizer makes different choices; setting `codegen-units = 1` makes this commit and its module-per-item alternative indistinguishable, and both land at the faster number.
Layout was ruled out separately, by permuting function order within an untouched file and remeasuring.

The branch was +301 lines against master-next and is now -67.
